Move Python SDK tests to their appropriate modules
